Centralize task types in `taskTypes.ts` and reuse across files
