+++ nextjs-task-system/types/taskTypes.ts
@@ -1,0 +1,13 @@
+export type TaskPriority = "low" | "medium" | "high";
+export type TaskStatus = "pending" | "inProgress" | "completed";
+
+export interface Task {
+  id: number;
+  title: string;
+  description: string;
+  status: TaskStatus;
+  dueDate: string;
+  assignedTo: string;
+  priority: TaskPriority;
+  comments: string[];
+}
